backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.6.1
starlette==0.36.3
PyJWT==2.8.0 
python-dotenv==1.0.1